Initial Changes For Pair Contracts
Introducing the token custom contract in the code for leveraging token and its interface for the Pair Contract.

The Following Files are being Added-

Allowance , Balance , Metadata  , Storage types , Mod and Total Supply and the Contract Files.

Next Steps

Introduce the Pair Contract Code For the Creation of the pair token contract s via factory and router contracts.